Refactor circle-of-fourths-with-inversions.
Pretty dang proud of this one NGL.

With this new expression, you write the chord form out exactly once. Then the operators take care of generating the circle of fourths and all inversions, even giving an opportunity to insert string transference where needed. It labels the rows flexibly and emits the entire thing in a nice HTML table.